+++ src/icons/animation/pauseIcon.ts
@@ -1,0 +1,5 @@
+import pause from './pause.svg?raw';
+export const icon = {
+    id: 'pause',
+    svg: pause
+}