+++ src/icons/animation/playIcon.ts
@@ -1,0 +1,5 @@
+import play from './play.svg?raw';
+export const icon = {
+    id: 'play',
+    svg: play
+}